Fix: add project 'paper', delete csv file from git
