--- server/paperflow-functions-v1/paperflowServer/paper/Real Error.csv
@@ -1,69 +1,0 @@
-﻿ID,PaperID,#Real Error
-1,1732412,4
-2,1732444,2
-3,1732472,52
-4,1732473,57
-5,1732525,1
-6,1732532,1
-7,1732536,2
-8,1732541,47
-9,1732577,2
-10,1732583,84
-11,1732622,16
-12,1732625,1
-13,1732628,9
-14,1732646,5
-15,1732647,1
-16,1732668,45
-17,1732702,15
-18,1732704,52
-19,1732711,35
-20,1732731,1
-21,1732744,53
-22,1732747,88
-23,1732772,112
-24,1732807,1
-25,1732854,305
-26,1732856,1
-27,1732936,80
-28,1732944,45
-29,1732950,1
-30,1732961,2
-31,1732967,1
-32,1732990,38
-33,1732993,25
-34,1732996,1
-35,1733000,1
-36,1733006,1
-37,1733050,4
-38,1733054,1
-39,1733089,12
-40,1733094,2
-41,1733102,1
-42,1733116,22
-43,1733144,604
-44,1733169,2
-45,1733170,171
-46,1733187,1
-47,1733209,1
-48,1733217,9
-49,1733237,84
-50,1733240,1
-51,1733243,94
-52,1733251,20
-53,1733252,12
-54,1733254,18
-55,1733256,1
-56,1733259,5
-57,1733261,1
-58,1733273,2
-59,1733293,2
-60,1733294,1
-61,1733301,25
-62,1733308,3
-63,1733311,2
-64,1733316,24
-65,1733317,5
-66,1733318,6
-67,1733319,2
-68,1733331,23
