+++ server/paperflow-functions-v1/paperflowServer/paper/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PaperConfig(AppConfig):
+    name = 'paper'
