+++ server/paperflow-functions-v1/paperflowServer/paper/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+# Create your views here.
+def index(request):
+    return HttpResponse("Hello, world! This is Paper Page!")
