+++ server/paperflow-functions-v1/paperflowServer/paperflowServer/settings.py
@@ -36,7 +36,7 @@ INSTALLED_APPS = [
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
-    'backendApp.apps.BackendappConfig'
+    'paper.apps.PaperConfig'
 ]
 
 MIDDLEWARE = [
